test(usage): isolate native history collectors
